整理

+++ sw2020_server/src/main/java/com/example/demo/domain/RecordMark.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class RecordMark{
 	String id;
-	int time; //标记的时间毫秒数
+	long time; //标记的时间毫秒数
 	String mark; //关注care 疑问doubt
 	String classId;//标注课堂的id
 	String userId;//标记者的id
